--- x-pack/plugin/ml/qa/native-multi-node-tests/src/javaRestTest/java/org/elasticsearch/xpack/ml/integration/CoordinatedInferenceActionIT.java
@@ -1,2 +1,0 @@
-package org.elasticsearch.xpack.ml.integration;public class CoordinatedInferenceActionIT {
-}
